refactor(check): extract utils.ts from checker.ts
Extract node classification and numeric utilities into a dedicated
utils.ts module as part of the checker modularization effort:

- Type guard: isValidExprResult
- Node classification: isTerminator, isStatementNode, isExpressionNode, isPatternNode
- Numeric utilities: INT_BOUNDS, fitsInBaseBounds, fitsInConstraints, splitBigIntTo32BitParts, isValidF32, isFloatType, isIntegerType
- Operator utilities: isIntegerOnlyOperator, isComparisonOperator, isLogicalOperator, getOperatorName
